+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSDraggingEnumerator.cs
@@ -1,5 +1,3 @@
-using Xamarin.Mac.System.Mac;
-
 namespace AppKit;
 
-public delegate void NSDraggingEnumerator(NSDraggingItem draggingItem, nint idx, ref bool stop);
+public delegate void NSDraggingEnumerator(NSDraggingItem draggingItem, int idx, ref bool stop);
