+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSImagePartialEventArgs.cs
@@ -1,4 +1,4 @@
-using Xamarin.Mac.System.Mac;
+using System;
 
 namespace AppKit;
 
@@ -6,9 +6,9 @@ public class NSImagePartialEventArgs : EventArgs
 {
 	public NSImageRep Rep { get; set; }
 
-	public nint Rows { get; set; }
+	public long Rows { get; set; }
 
-	public NSImagePartialEventArgs(NSImageRep rep, nint rows)
+	public NSImagePartialEventArgs(NSImageRep rep, long rows)
 	{
 		Rep = rep;
 		Rows = rows;
