+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSRuleEditorNumberOfChildren.cs
@@ -1,6 +1,5 @@
 using Foundation;
-using Xamarin.Mac.System.Mac;
 
 namespace AppKit;
 
-public delegate nint NSRuleEditorNumberOfChildren(NSRuleEditor editor, NSObject criterion, NSRuleEditorRowType rowType);
+public delegate long NSRuleEditorNumberOfChildren(NSRuleEditor editor, NSObject criterion, NSRuleEditorRowType rowType);
