+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSRulerEditorChildCriterion.cs
@@ -1,6 +1,5 @@
 using Foundation;
-using Xamarin.Mac.System.Mac;
 
 namespace AppKit;
 
-public delegate NSObject NSRulerEditorChildCriterion(NSRuleEditor editor, nint index, NSObject criterion, NSRuleEditorRowType rowType);
+public delegate NSObject NSRulerEditorChildCriterion(NSRuleEditor editor, long index, NSObject criterion, NSRuleEditorRowType rowType);
