+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSRulerEditorDisplayValue.cs
@@ -1,6 +1,5 @@
 using Foundation;
-using Xamarin.Mac.System.Mac;
 
 namespace AppKit;
 
-public delegate NSObject NSRulerEditorDisplayValue(NSRuleEditor editor, NSObject criterion, nint row);
+public delegate NSObject NSRulerEditorDisplayValue(NSRuleEditor editor, NSObject criterion, long row);
