+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSRulerEditorPredicateParts.cs
@@ -1,6 +1,5 @@
 using Foundation;
-using Xamarin.Mac.System.Mac;
 
 namespace AppKit;
 
-public delegate NSDictionary NSRulerEditorPredicateParts(NSRuleEditor editor, NSObject criterion, NSObject value, nint row);
+public delegate NSDictionary NSRulerEditorPredicateParts(NSRuleEditor editor, NSObject criterion, NSObject value, long row);
