+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSSavePanelComplete.cs
@@ -1,5 +1,3 @@
-using Xamarin.Mac.System.Mac;
-
 namespace AppKit;
 
-public delegate void NSSavePanelComplete(nint result);
+public delegate void NSSavePanelComplete(int result);
