+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTableReorder.cs
@@ -1,5 +1,3 @@
-using Xamarin.Mac.System.Mac;
-
 namespace AppKit;
 
-public delegate bool NSTableReorder(NSTableView tableView, nint columnIndex, nint newColumnIndex);
+public delegate bool NSTableReorder(NSTableView tableView, long columnIndex, long newColumnIndex);
