+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTableViewCell.cs
@@ -1,5 +1,3 @@
-using Xamarin.Mac.System.Mac;
-
 namespace AppKit;
 
-public delegate bool NSTableViewCell(NSTableView tableView, NSCell cell, NSTableColumn tableColumn, nint row);
+public delegate bool NSTableViewCell(NSTableView tableView, NSCell cell, NSTableColumn tableColumn, long row);
