+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTableViewCellGetter.cs
@@ -1,5 +1,3 @@
-using Xamarin.Mac.System.Mac;
-
 namespace AppKit;
 
-public delegate NSCell NSTableViewCellGetter(NSTableView tableView, NSTableColumn tableColumn, nint row);
+public delegate NSCell NSTableViewCellGetter(NSTableView tableView, NSTableColumn tableColumn, long row);
