+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTableViewColumnRowString.cs
@@ -1,5 +1,3 @@
-using Xamarin.Mac.System.Mac;
-
 namespace AppKit;
 
-public delegate string NSTableViewColumnRowString(NSTableView tableView, NSTableColumn tableColumn, nint row);
+public delegate string NSTableViewColumnRowString(NSTableView tableView, NSTableColumn tableColumn, long row);
