+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTableViewRowGetter.cs
@@ -1,5 +1,3 @@
-using Xamarin.Mac.System.Mac;
-
 namespace AppKit;
 
-public delegate NSTableRowView NSTableViewRowGetter(NSTableView tableView, nint row);
+public delegate NSTableRowView NSTableViewRowGetter(NSTableView tableView, long row);
