+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTableViewRowHandler.cs
@@ -1,5 +1,3 @@
-using Xamarin.Mac.System.Mac;
-
 namespace AppKit;
 
-public delegate void NSTableViewRowHandler(NSTableRowView rowView, nint row);
+public delegate void NSTableViewRowHandler(NSTableRowView rowView, int row);
