+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTextViewCellPasteboard.cs
@@ -1,5 +1,3 @@
-using Xamarin.Mac.System.Mac;
-
 namespace AppKit;
 
-public delegate bool NSTextViewCellPasteboard(NSTextView view, NSTextAttachmentCell cell, nuint charIndex, NSPasteboard pboard, string type);
+public delegate bool NSTextViewCellPasteboard(NSTextView view, NSTextAttachmentCell cell, uint charIndex, NSPasteboard pboard, string type);
