+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTextViewCellPosition.cs
@@ -1,5 +1,3 @@
-using Xamarin.Mac.System.Mac;
-
 namespace AppKit;
 
-public delegate string[] NSTextViewCellPosition(NSTextView view, NSTextAttachmentCell forCell, nuint charIndex);
+public delegate string[] NSTextViewCellPosition(NSTextView view, NSTextAttachmentCell forCell, ulong charIndex);
