+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTextViewEventMenu.cs
@@ -1,5 +1,3 @@
-using Xamarin.Mac.System.Mac;
-
 namespace AppKit;
 
-public delegate NSMenu NSTextViewEventMenu(NSTextView view, NSMenu menu, NSEvent theEvent, nuint charIndex);
+public delegate NSMenu NSTextViewEventMenu(NSTextView view, NSMenu menu, NSEvent theEvent, ulong charIndex);
