+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTextViewLink.cs
@@ -1,6 +1,5 @@
 using Foundation;
-using Xamarin.Mac.System.Mac;
 
 namespace AppKit;
 
-public delegate bool NSTextViewLink(NSTextView textView, NSObject link, nuint charIndex);
+public delegate bool NSTextViewLink(NSTextView textView, NSObject link, ulong charIndex);
