+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTextViewTextChecked.cs
@@ -1,6 +1,5 @@
 using Foundation;
-using Xamarin.Mac.System.Mac;
 
 namespace AppKit;
 
-public delegate NSTextCheckingResult[] NSTextViewTextChecked(NSTextView view, NSRange range, NSTextCheckingTypes checkingTypes, NSDictionary options, NSTextCheckingResult[] results, NSOrthography orthography, nint wordCount);
+public delegate NSTextCheckingResult[] NSTextViewTextChecked(NSTextView view, NSRange range, NSTextCheckingTypes checkingTypes, NSDictionary options, NSTextCheckingResult[] results, NSOrthography orthography, long wordCount);
