+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTextViewTooltip.cs
@@ -1,5 +1,3 @@
-using Xamarin.Mac.System.Mac;
-
 namespace AppKit;
 
-public delegate string NSTextViewTooltip(NSTextView textView, string tooltip, nuint characterIndex);
+public delegate string NSTextViewTooltip(NSTextView textView, string tooltip, ulong characterIndex);
