+++ Source/Platform/Mac/Xamarin.Mac/ImageKit/IKImageBrowserViewIndexEventEventArgs.cs
@@ -1,15 +1,15 @@
+using System;
 using AppKit;
-using Xamarin.Mac.System.Mac;
 
 namespace ImageKit;
 
 public class IKImageBrowserViewIndexEventEventArgs : EventArgs
 {
-	public nint Index { get; set; }
+	public int Index { get; set; }
 
 	public NSEvent Nsevent { get; set; }
 
-	public IKImageBrowserViewIndexEventEventArgs(nint index, NSEvent nsevent)
+	public IKImageBrowserViewIndexEventEventArgs(int index, NSEvent nsevent)
 	{
 		Index = index;
 		Nsevent = nsevent;
